This is a rescue repository for the MOOSE project. It originally was at revision 8. I lost the repository when upgrading OSs, as I did not back up the /opt directory. Fortunately I was able to recover all the dated releases in the tags directory. SVN maintains a local copy of the most recent revision. I recreated that thanks to advice from Duncan Murdoch on the Subversion users mailing group. To summarize: this initial checkin corresponds to the earlier revision 8 of MOOSE, plus all the tagged releases.
-- Upi Bhalla, June 11 2006.

+++ scheduling/ClockTickWrapper.cpp
@@ -94,7 +94,7 @@ const Cinfo ClockTickWrapper::cinfo_(
 void ClockTickWrapper::innerSetPath( const string& path )
 {
 	path_ = path;
-	size_t pos = path.find_last_of("/");
+	unsigned int pos = path.find_last_of("/");
 	if ( pos == string::npos || pos == path.length()) {
 		cerr << "Error:ClockTickWrapper::innerSetPath: no finfo name in" << path << "\n"; 
 		return;
